requirements.txt
bs4
requests
wikipedia
Python_Levenshtein
